moved BatchCollection from timing to measurement
