+++ enlighten/measurement/BatchCollectionFeature.py
